requirements.txt
langchain
langchain-core
langchain-community
langchain-google-genai
google-generativeai
huggingface-hub
faiss-cpu
sentence-transformers
python-dotenv
typer[all]